% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{BasicSeuratProcessingGdata}
\alias{BasicSeuratProcessingGdata}
\title{Helper function to run basic Seurat initial processing}
\usage{
BasicSeuratProcessingGdata(gdata)
}
\arguments{
\item{gdata}{A Seurat object}
}
\value{
A Seurat object after PCA, UMAP and leiden clustering
}
\description{
Helper function to run basic Seurat initial processing
}
