% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{compute_gene_classes}
\alias{compute_gene_classes}
\title{Helper function: Computes what type category each gene is in the network.
In particular, finds the irreducible subset (assuming there is only one!)}
\usage{
compute_gene_classes(nandonet)
}
\arguments{
\item{nandonet}{One network}
}
\description{
Helper function: Computes what type category each gene is in the network.
In particular, finds the irreducible subset (assuming there is only one!)
}
