% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ComputeGeneScoreGO}
\alias{ComputeGeneScoreGO}
\title{Perform network-informed GO analysis. Requires global variables properly set up before}
\usage{
ComputeGeneScoreGO(
  allscores,
  GO.type = "MF",
  num_significant = 1000,
  mapping = "org.Hs.eg.db"
)
}
\arguments{
\item{allscores}{Data.frame with columns (netname, gene, score)}

\item{num_significant}{How many top genes to include for GO analysis}

\item{mapping}{Which gene database to use}
}
\description{
Perform network-informed GO analysis. Requires global variables properly set up before
}
