% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ComputeGOFromHP}
\alias{ComputeGOFromHP}
\title{Compute function of gene using GO over hitting probablities, i.e, which genes are regulated
by this gene?}
\usage{
ComputeGOFromHP(
  nandonets,
  gene,
  reference_network,
  GO.type = c("MF", "BP", "CC"),
  num_significant = 1000,
  mapping = "org.Hs.eg.db"
)
}
\arguments{
\item{gene}{Gene to consider}

\item{reference_network}{The null distribution; one sensible choice is a network consisting of all cells}

\item{num_significant}{How many top genes to include for GO analysis}

\item{mapping}{Which gene database to use}

\item{nandonet}{ListOfNandoNetwork}
}
\value{
Data.frame to be plotted with PlotGeneGO
}
\description{
Compute function of gene using GO over hitting probablities, i.e, which genes are regulated
by this gene?
}
