% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ComputeHittingProbability.NandoNetwork}
\alias{ComputeHittingProbability.NandoNetwork}
\alias{ComputeHittingProbability.ListOfNandoNetwork}
\title{Compute hitting probabilities for a network.}
\usage{
\method{ComputeHittingProbability}{ListOfNandoNetwork}(nandonet)

\method{ComputeHittingProbability}{ListOfNandoNetwork}(nandonets)
}
\arguments{
\item{nandonet}{One NandoNetwork}

\item{nandonets}{A ListOfNandoNetwork}
}
\value{
A NandoNetwork object with hitting probability set

A ListOfNandoNetwork object with hitting probability set
}
\description{
A crucial trick is used internally to do this fast. HPs are first calculated for TFs alone, as
this requires a full linear equation to work out. Then nonTFs are calculated, and since these are
unreachable, their HPs can be worked out by simply weighing HPs from the TFs
}
