% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ComputeHittingProbabilityFromSS}
\alias{ComputeHittingProbabilityFromSS}
\title{For each network, compute hitting probability, starting from steady state}
\usage{
ComputeHittingProbabilityFromSS(nandonets)
}
\arguments{
\item{nandonets}{A ListOfNandoNetworks}
}
\value{
A sparse matrix of hitting probabilities; each row is one network
}
\description{
For each network, compute hitting probability, starting from steady state
}
