% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ComputeNandoNetworkSimilarity}
\alias{ComputeNandoNetworkSimilarity}
\title{Compute network similarities}
\usage{
ComputeNandoNetworkSimilarity(
  nandonets,
  method = c("jaccard", "ssi"),
  datarep = c("mc", "ss", "hp", "hpss")
)
}
\arguments{
\item{nandonets}{ListOfNandoNetwork}

\item{method}{The similarity metric to apply}
}
\value{
Matrix of similarity scores
}
\description{
Compute network similarities
}
\details{
Similarities done using fuzzy logic, as our edges have weights \link{0,1} (if an edge not in list then p=0)
https://en.wikipedia.org/wiki/Fuzzy_set_operations
discussion of |A| here: https://gvpress.com/journals/IJEIC/vol4_no1/2.pdf

Jaccard index: https://en.wikipedia.org/wiki/Jaccard_index
|A intersect B| / |A union B|

SSI: https://en.wikipedia.org/wiki/Overlap_coefficient
|A intersect B| / min(|A|, |B|)
}
