% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ComputeNandoWalktrap.NandoNetwork}
\alias{ComputeNandoWalktrap.NandoNetwork}
\alias{ComputeNandoWalktrap.ListOfNandoNetwork}
\title{Compute a series of P^t to figure out about the neighborhood of genes, for one network}
\usage{
\method{ComputeNandoWalktrap}{NandoNetwork}(nandonet, num_mul, only_irreducible_TF = TRUE)

\method{ComputeNandoWalktrap}{ListOfNandoNetwork}(nandonets, num_mul, only_irreducible_TF = TRUE)
}
\arguments{
\item{nandonet}{One network}

\item{num_mul}{Calculate for P^1 ... P^num_mul}

\item{only_irreducible_TF}{Should only irreducible TFs be kept, and other TFs removed?}

\item{nandonets}{List of network}
}
\value{
A NandoWalktrap object

A ListOfNandoWalktrap object
}
\description{
Compute a series of P^t to figure out about the neighborhood of genes, for one network

Compute a series of P^t to figure out about the neighborhood of genes, for a list of networks
}
