% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ComputeProbabilitySimilarity}
\alias{ComputeProbabilitySimilarity}
\title{Comparison of distributions. For the definition of Jaccard and SSI, see
documentation of ComputeNandoNetworkSimilarity}
\usage{
ComputeProbabilitySimilarity(tocompare, method = c("jaccard", "ssi"))
}
\arguments{
\item{tocompare}{A matrix of probabilities, with each sample being a column}

\item{method}{The similarity metric to apply}
}
\value{
Matrix of similarity scores
}
\description{
Comparison of distributions. For the definition of Jaccard and SSI, see
documentation of ComputeNandoNetworkSimilarity
}
