% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ComputeSimplifiedMatrix.ListOfNandoNetwork}
\alias{ComputeSimplifiedMatrix.ListOfNandoNetwork}
\title{TODO just cannot make this one work. not clue why...  update: now maybe working. to test}
\usage{
\method{ComputeSimplifiedMatrix}{ListOfNandoNetwork}(nandonets, keep_genes)
}
\arguments{
\item{nandonets}{ListOfNandoNetwork}

\item{keep_genes}{Genes to retain in the network}
}
\value{
List of transition matrices
}
\description{
Compute an equivalent transition matrix where only a subset of genes are kept.
This currently only works over the set of irreducible TFs.
The output can be used to visualize simplified GRNs
}
