% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ComputeSimplifiedMatrix.NandoNetwork}
\alias{ComputeSimplifiedMatrix.NandoNetwork}
\title{Compute an equivalent transition matrix where only a subset of genes are kept.
This currently only works over the set of irreducible TFs.
The output can be used to visualize simplified GRNs}
\usage{
\method{ComputeSimplifedMatrix}{NandoNetwork}(net, keep_genes)
}
\arguments{
\item{net}{NandoNetwork}

\item{keep_genes}{Genes to retain in the network}
}
\value{
A transition matrix
}
\description{
Compute an equivalent transition matrix where only a subset of genes are kept.
This currently only works over the set of irreducible TFs.
The output can be used to visualize simplified GRNs
}
