% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ComputeSteadyState.NandoNetwork}
\alias{ComputeSteadyState.NandoNetwork}
\alias{ComputeSteadyState.ListOfNandoNetwork}
\title{Compute steady state for a network. Only the irreducible TFs will be considered}
\usage{
\method{ComputeSteadyState}{NandoNetwork}(nandonet)

\method{ComputeSteadyState}{ListOfNandoNetwork}(nandonets)
}
\arguments{
\item{nandonet}{One network}

\item{nandonets}{A list of networks}
}
\value{
A NandoNetwork object with steady state set

A ListOfNandoNetwork object with steady state set
}
\description{
Compute steady state for a network. Only the irreducible TFs will be considered

Compute steady state for each network.
This function is able to parallelize using foreach if set up properly
}
