% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ComputeSteadyStateChangeEdgeflow}
\alias{ComputeSteadyStateChangeEdgeflow}
\title{In moving from one state state pi_i to pi_j, which edges does the probability
mainly flow over?}
\usage{
ComputeSteadyStateChangeEdgeflow(netA, netB)
}
\arguments{
\item{netA, netB}{Two NandoNetworks to be compared}
}
\value{
Matrix, from each gene showing how much flowed to another
}
\description{
In moving from one state state pi_i to pi_j, which edges does the probability
mainly flow over?
}
