% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{CreateOneNandoNetwork}
\alias{CreateOneNandoNetwork}
\title{Helper function: Prepare one network from a long list of SHAP.
Sets up gene categories}
\usage{
CreateOneNandoNetwork(shaps)
}
\arguments{
\item{shaps}{Long format of shap scores}
}
\description{
Helper function: Prepare one network from a long list of SHAP.
Sets up gene categories
}
