% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyexport.R
\name{ExportGeneCategories}
\alias{ExportGeneCategories}
\title{Store category of each gene}
\usage{
ExportGeneCategories(nando_dir, nandonets)
}
\arguments{
\item{nando_dir}{Nando directory}

\item{nandonets}{ListOfNandoNetwork}
}
\value{
Nothing
}
\description{
Store category of each gene
}
