% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyexport.R
\name{ExportShapSqlite}
\alias{ExportShapSqlite}
\title{Store SHAPs per network, gene and region}
\usage{
ExportShapSqlite(nando_dir)
}
\value{
Nothing
}
\description{
Store SHAPs per network, gene and region
}
