% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyexport.R
\name{ExportShinyNando.HP}
\alias{ExportShinyNando.HP}
\title{Export HP data available to ShinyNando}
\usage{
ExportShinyNando.HP(nando_dir, nandonets)
}
\arguments{
\item{nandonets}{A ListOfNandoNetwork}
}
\value{
Nothing
}
\description{
Export HP data available to ShinyNando
}
