% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyexport.R
\name{ExportShinyNando.HP.SS}
\alias{ExportShinyNando.HP.SS}
\title{Export HP.SS data available to ShinyNando}
\usage{
ExportShinyNando.HP.SS(nando_dir, nandonets)
}
\arguments{
\item{nandonets}{A ListOfNandoNetwork}
}
\value{
Nothing
}
\description{
Export HP.SS data available to ShinyNando
}
