% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ExportTransitionMatrixCSV}
\alias{ExportTransitionMatrixCSV}
\title{Export transition matrix to a CSV file for e.g. reading in cytoscape}
\usage{
ExportTransitionMatrixCSV(tmat, fname)
}
\arguments{
\item{tmat}{A transition matrix}

\item{fname}{Name of a csv file to write to}
}
\description{
Export transition matrix to a CSV file for e.g. reading in cytoscape
}
