% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{fast_entropy_sparsematrix}
\alias{fast_entropy_sparsematrix}
\title{Helper function: Compute entropy over a transition matrix}
\usage{
fast_entropy_sparsematrix(mat)
}
\arguments{
\item{mat}{Matrix, which must be in format of TsparseMatrix}
}
\value{
One-column matrix with the entropy from each gene
}
\description{
Helper function: Compute entropy over a transition matrix
}
