% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{GeneCategories.ListOfNandoNetwork}
\alias{GeneCategories.ListOfNandoNetwork}
\title{Get what type each gene is}
\usage{
GeneCategories.ListOfNandoNetwork(nandonets)
}
\arguments{
\item{net}{A NandoNetwork}
}
\value{
A data.frame with one column per network
}
\description{
Get what type each gene is
}
