% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{GeneCategories.NandoNetwork}
\alias{GeneCategories.NandoNetwork}
\title{Get what type each gene is}
\usage{
GeneCategories.NandoNetwork(net)
}
\arguments{
\item{net}{A NandoNetwork}
}
\value{
A data.frame with two columns, gene and category
}
\description{
Get what type each gene is
}
