% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{GeneCategoriesDF}
\alias{GeneCategoriesDF}
\title{Return what type of gene each of them is}
\usage{
GeneCategoriesDF(net)
}
\arguments{
\item{net}{A NandoNetwork}
}
\value{
A data.frame with two columns, gene and category
}
\description{
Return what type of gene each of them is
}
