% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{GetHittingProbabilityTo.ListOfNandoNetwork}
\alias{GetHittingProbabilityTo.ListOfNandoNetwork}
\title{Get hitting probabilities for all genes to a particular gene}
\usage{
GetHittingProbabilityTo.ListOfNandoNetwork(nandonets, to_gene)
}
\arguments{
\item{nandonets}{ListOfNandoNetwork}

\item{to_gene}{To which gene}
}
\description{
Get hitting probabilities for all genes to a particular gene
}
