% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{GetWalktrapEntropy}
\alias{GetWalktrapEntropy}
\title{Gather entropies across all networks for chosen genes and steps}
\usage{
GetWalktrapEntropy(walktraps, keep_genes = NULL, keep_steps = NULL)
}
\arguments{
\item{walktraps}{A ListOfNandoWalktraps}

\item{keep_genes}{Which genes to extract. Default is all (which is a lot!)}

\item{keep_steps}{Which steps to keep. Default is all}
}
\value{
Molten data.frame with all values
}
\description{
Gather entropies across all networks for chosen genes and steps
}
