% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{HittingProbabilitiesToSeurat}
\alias{HittingProbabilitiesToSeurat}
\title{Convert hitting probabilities into Seurat to explore the cell state space using common single cell tools.}
\usage{
HittingProbabilitiesToSeurat(
  nandonets,
  keep_genes_from = NULL,
  keep_genes_to = NULL
)
}
\arguments{
\item{walktraps}{A ListOfNandoNetwork}

\item{keep_genes}{Which genes to subset by; default is to keep all}
}
\value{
A Seurat object
}
\description{
Convert hitting probabilities into Seurat to explore the cell state space using common single cell tools.
}
