% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyexport.R
\name{ImportGeneCategories}
\alias{ImportGeneCategories}
\title{Read category of each gene}
\usage{
ImportGeneCategories(nando_dir)
}
\arguments{
\item{nando_dir}{Nando directory}
}
\value{
Data frame with categories per net
}
\description{
Read category of each gene
}
