% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyexport.R
\name{ImportShapsForGene}
\alias{ImportShapsForGene}
\title{Get all the upstream SHAP info for one gene, including region}
\usage{
ImportShapsForGene(nando_dir, genename)
}
\description{
Get all the upstream SHAP info for one gene, including region
}
