% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyexport.R
\name{ImportShinyNando.HP.SS}
\alias{ImportShinyNando.HP.SS}
\title{Import SS data available to ShinyNando}
\usage{
ImportShinyNando.HP.SS(nando_dir)
}
\arguments{
\item{nandonets}{A ListOfNandoNetwork}
}
\value{
Nothing
}
\description{
Import SS data available to ShinyNando
}
