% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{ListOfNandoWalktrap-class}
\alias{ListOfNandoWalktrap-class}
\title{An S4 class to represent a list of walktrap analyses}
\description{
An S4 class to represent a list of walktrap analyses
}
\section{Slots}{

\describe{
\item{\code{nets}}{All walktraps}
}}

