% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{LoadNandoNetworks}
\alias{LoadNandoNetworks}
\title{Load networks previously summarized by cluster}
\usage{
LoadNandoNetworks(nando_dir, keep_clusters = NULL)
}
\arguments{
\item{nando_dir}{The directory holding per-cluster SHAP scores}

\item{keep_clusters}{Optional list of clusters to extract; default is all of them}
}
\value{
A ListofNandoNetwork object
}
\description{
Load networks previously summarized by cluster
}
