% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{NandoNetwork-class}
\alias{NandoNetwork-class}
\alias{NandoNetwork}
\title{An S4 class to represent the model of a network.}
\description{
Note that if a gene is never listed in the TF column, it will be considered a nonTF.
}
\section{Slots}{

\describe{
\item{\code{list_all_gene}}{All genes (TFs and nonTFs)}

\item{\code{list_tfs}}{All TFs every mentioned}

\item{\code{list_irreducible}}{TFs that are members of the irreducible net}

\item{\code{list_nontfs}}{All nonTFs}
}}

