% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{NandoWalktrap-class}
\alias{NandoWalktrap-class}
\title{An S4 class to represent a walktrap analysis.
It holds P^t and entropy for each step}
\description{
An S4 class to represent a walktrap analysis.
It holds P^t and entropy for each step
}
\section{Slots}{

\describe{
\item{\code{mat}}{All P stored concatenated}

\item{\code{meta}}{For each row of P, what it corresponds to}

\item{\code{entropy}}{Entropy for every step}
}}

