% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{NandoWalktrapToSeurat}
\alias{NandoWalktrapToSeurat}
\title{Convert Walktrap into Seurat to explore the cell state space using common single cell tools.}
\usage{
NandoWalktrapToSeurat(walktraps, keep_steps = NULL, keep_genes = NULL)
}
\arguments{
\item{walktraps}{A ListOfNandoWalktraps}

\item{keep_steps}{Which steps to subset by; default is to keep all}

\item{keep_genes}{Which genes to subset by; default is to keep all}
}
\value{
A Seurat object
}
\description{
Convert Walktrap into Seurat to explore the cell state space using common single cell tools.
}
