% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{PlotGeneGOvsNetworks}
\alias{PlotGeneGOvsNetworks}
\title{Plot the result of network-informed GSEA as a heatmap, celltype vs GO term}
\usage{
PlotGeneGOvsNetworks(go_result, padj_cutoff = 0.1)
}
\arguments{
\item{go_result}{Data.frame with columns (netname, pathway, padj, log_padj)}
}
\value{
A gglot2 object
}
\description{
Plot the result of network-informed GSEA as a heatmap, celltype vs GO term
}
