% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{PlotTopProbabilityMatrix}
\alias{PlotTopProbabilityMatrix}
\title{Plot a matrix consisting of probabilities}
\usage{
PlotTopProbabilityMatrix(probs, min.pmean = 0.01, dolog = TRUE)
}
\arguments{
\item{probs}{A matrix of probabilities; can be sparse or dense}

\item{min.pmean}{A lower cutoff on p values}

\item{dolog}{Show log10 p values or not}
}
\value{
A ggplot2 object
}
\description{
Plot a matrix consisting of probabilities
}
