% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{PlotWalktrapEntropy}
\alias{PlotWalktrapEntropy}
\title{Gather and plot entropy for chosen genes over time}
\usage{
PlotWalktrapEntropy(for_genes)
}
\arguments{
\item{for_genes}{Genes to show}
}
\value{
A ggplot2 object
}
\description{
Gather and plot entropy for chosen genes over time
}
