% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{PrepareNandoDir}
\alias{PrepareNandoDir}
\title{Prepare seurat object for analysis with Pando, and then Nando}
\usage{
PrepareNandoDir(
  nando_dir,
  adata,
  include_genes = NULL,
  use_genome = BSgenome.Hsapiens.UCSC.hg38
)
}
\arguments{
\item{nando_dir}{Location of all Nando files. Created if not existing}

\item{adata}{Seurat object}

\item{include_genes}{List of genes to fit; using highly variable genes if not set}

\item{use_genome}{The genome to use for the putative network}
}
\value{
Nothing, creates directory and initial files
}
\description{
Prepare seurat object for analysis with Pando, and then Nando
}
