% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyexport.R
\name{PrepareProbabilityMatrixH5}
\alias{PrepareProbabilityMatrixH5}
\title{Prepare a set of stored matrices for reading.
Read essentials to speed up later access}
\usage{
PrepareProbabilityMatrixH5(fname)
}
\arguments{
\item{fname}{Name of the file}
}
\description{
Prepare a set of stored matrices for reading.
Read essentials to speed up later access
}
