% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{ProbabilityMatrixH5-class}
\alias{ProbabilityMatrixH5-class}
\title{An S4 class to represent a probability matrix file.
Provides speedy out-of-memory access.}
\description{
An S4 class to represent a probability matrix file.
Provides speedy out-of-memory access.
}
\section{Slots}{

\describe{
\item{\code{fname}}{Name of the file}

\item{\code{totalcols}}{All the column names ever mentioned}

\item{\code{totalrows}}{All the row names ever mentioned}

\item{\code{list_cols}}{Column names of each matrix}

\item{\code{list_rows}}{Row names of each matrix}

\item{\code{list_names}}{Names of each network}
}}

