% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyexport.R
\name{ReadProbabilityMatricesH5ForGene}
\alias{ReadProbabilityMatricesH5ForGene}
\title{Read probability matrices for one gene}
\usage{
ReadProbabilityMatricesH5ForGene(matrixfile, netnames, genename)
}
\description{
Read probability matrices for one gene
}
