% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{RunPandoParallel}
\alias{RunPandoParallel}
\title{Run Pando to fit genes and store SHAP scores per cell}
\usage{
RunPandoParallel(nando_dir)
}
\arguments{
\item{nando_dir}{Location of all Nando files}
}
\value{
Nothing, saves to disk immediately
}
\description{
This function should be used in an R script run in a SLURM batch array.
It will use SLURM_ARRAY_TASK_COUNT and SLURM_ARRAY_TASK_ID to divide the
summarization between different nodes
}
