% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ScoreProbabilityDifferences}
\alias{ScoreProbabilityDifferences}
\title{Scores differences between networks. SetNandoNetworkReplicateTypes must have
been called first to set up network type and replicate names.
Networks of type "" will be excluded.}
\usage{
ScoreProbabilityDifferences(nandonets, probs, reference_network)
}
\arguments{
\item{probs}{Data.frame having columns (netname, gene, p)}

\item{reference_network}{Name of network type to compare all other network types to}

\item{nanonets}{ListOfNandoNetwork}
}
\value{
Data.frame with columns (gene, score)
}
\description{
Scores differences between networks. SetNandoNetworkReplicateTypes must have
been called first to set up network type and replicate names.
Networks of type "" will be excluded.
}
