% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{SelectClusteringForNando}
\alias{SelectClusteringForNando}
\title{Selection of clusterings over which SHAPs will be summarized}
\usage{
SelectClusteringForNando(nando_dir, use_clusterings)
}
\arguments{
\item{use_clusterings}{Data frame, each column with a clustering. Row names must be all the cell names from the Seurat object}
}
\description{
Selection of clusterings over which SHAPs will be summarized
}
