% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{SetNandoNetworkReplicateTypes}
\alias{SetNandoNetworkReplicateTypes}
\title{Set the types of replicate networks. Needed for some statistical testing.
Either fit networks for each biological replicate, or perform random
subsetting for at least a technical variance analysis.
The arguments separate the original network name into (replicate_name, replicate_id)}
\usage{
SetNandoNetworkReplicateTypes(nandonets, replicate_type, replicate_id)
}
\arguments{
\item{nandonets}{ListOfNandoNetwork}

\item{replicate_type}{Name of replicate type, same length as number of networks}

\item{replicate_id}{ID of the source i.e. donor ID}
}
\description{
Set the types of replicate networks. Needed for some statistical testing.
Either fit networks for each biological replicate, or perform random
subsetting for at least a technical variance analysis.
The arguments separate the original network name into (replicate_name, replicate_id)
}
