% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{SteadyStateMatrix}
\alias{SteadyStateMatrix}
\title{For each network, extract steady states. Return as a matrix}
\usage{
SteadyStateMatrix(nandonets)
}
\arguments{
\item{nandonets}{A ListOfNandoNetworks}
}
\value{
Matrix of steady states; each row is one network
}
\description{
For each network, extract steady states. Return as a matrix
}
