% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyexport.R
\name{StoreProbabilityMatricesH5}
\alias{StoreProbabilityMatricesH5}
\title{Store list of matrices}
\usage{
StoreProbabilityMatricesH5(fname, list_matrix)
}
\value{
Nothing
}
\description{
Store list of matrices
}
