% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{SumListMatrices}
\alias{SumListMatrices}
\title{Output the sum of all matrices, with dimensions adapted to have rows and columns
of each matrix}
\usage{
SumListMatrices(listmat)
}
\description{
Output the sum of all matrices, with dimensions adapted to have rows and columns
of each matrix
}
