% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{SummarizeShapByCell}
\alias{SummarizeShapByCell}
\title{Compute summarized SHAP values for each TF in each cell.}
\usage{
SummarizeShapByCell(
  nando_dir,
  shap_dir,
  list_files_to_process,
  summarize_shap_for_cluster,
  task_id
)
}
\arguments{
\item{nando_dir}{The directory holding per-cluster SHAP scores}

\item{shap_dir}{Directory with SHAP per cell}

\item{list_files_to_process}{List of files to process}

\item{task_id}{Process ID when multiprocessing, or 0}
}
\value{
Nothing
}
\description{
This function is not meant to be used directly; most users
should rather use a wrapper that loops over all files
}
