% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{SummarizeShapByCluster}
\alias{SummarizeShapByCluster}
\title{Compute summarized SHAP values for a number of clusters.
Note that cells can be part of more than one cluster.}
\usage{
SummarizeShapByCluster(
  nando_dir,
  shap_dir,
  list_files_to_process,
  summarize_shap_for_cluster,
  taskid
)
}
\arguments{
\item{shap_dir}{Directory with SHAP per cell}

\item{list_files_to_process}{List of files to process}

\item{summarize_shap_for_cluster}{Clusterings of cells}

\item{taskid}{Process ID when multiprocessing, or 0}
}
\description{
This function is not meant to be used directly; most users should rather use a wrapper
that loops over all files
}
