% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{SummarizeShapByCluster_summarizeClusterNoreg}
\alias{SummarizeShapByCluster_summarizeClusterNoreg}
\title{Helper function}
\usage{
SummarizeShapByCluster_summarizeClusterNoreg(
  shap,
  for_category,
  onecat,
  curgene
)
}
\description{
Helper function
}
