% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{SummarizeShapByCluster_summarizeClusterReg}
\alias{SummarizeShapByCluster_summarizeClusterReg}
\title{Helper function}
\usage{
SummarizeShapByCluster_summarizeClusterReg(shap, for_category, onecat, curgene)
}
\description{
Helper function
}
