% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{SummarizeShapByClusterParallel}
\alias{SummarizeShapByClusterParallel}
\title{Compute summarized SHAP values for a number of clusters.
Note that cells can be part of more than one cluster.}
\usage{
SummarizeShapByClusterParallel(nando_dir)
}
\arguments{
\item{nando_dir}{Location of all Nando files}
}
\value{
Nothing
}
\description{
This function should be used in an R script run in a SLURM batch array.
It will use SLURM_ARRAY_TASK_COUNT and SLURM_ARRAY_TASK_ID to divide the
summarization between different nodes
}
