% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{TransitionMatrix.data.frame}
\alias{TransitionMatrix.data.frame}
\title{Make a transition matrix from SHAP scores. Primarily for internal use
but can be used for various subsets of the SHAP entries in a network}
\usage{
\method{TransitionMatrix}{data.frame}(
  onegrn,
  subsetGene = NULL,
  subsetTF = NULL,
  includegenes = c()
)
}
\arguments{
\item{onegrn}{Long format data frame of SHAPs}

\item{subsetGene, subsetTF}{Subset of genes to include. Default is all}

\item{includegenes}{Names of genes to forcefully add as row/columns in the matrix}
}
\value{
Transition matrix, in sparse matrix format to save memory
}
\description{
Make a transition matrix from SHAP scores. Primarily for internal use
but can be used for various subsets of the SHAP entries in a network
}
