% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{transitionmatrix_to_mc}
\alias{transitionmatrix_to_mc}
\title{Helper function: Convert a sparse transition matrix into one for the R markovchain library}
\usage{
transitionmatrix_to_mc(mat)
}
\arguments{
\item{mat}{Transition matrix in sparse format}
}
\value{
A markovchain object
}
\description{
Helper function: Convert a sparse transition matrix into one for the R markovchain library
}
