% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{TransitionMatrixToIgraph}
\alias{TransitionMatrixToIgraph}
\title{Convert a transition matrix into a weighted directed igraph object}
\usage{
TransitionMatrixToIgraph(tmat)
}
\arguments{
\item{tmat}{Transition matrix in sparse format}
}
\value{
igraph object
}
\description{
Convert a transition matrix into a weighted directed igraph object
}
