% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ComputeGOFromHP}
\alias{ComputeGOFromHP}
\title{Compute function of gene using GO over hitting probablities, i.e, which genes are regulated
by this gene?}
\usage{
ComputeGOFromHP(nandonets, gene, reference_network = "ALL")
}
\arguments{
\item{gene}{Gene to consider}

\item{reference_network}{The null distribution; one sensible choice is a network consisting of all cells}

\item{nandonet}{ListOfNandoNetwork}
}
\value{
Data.frame to be plotted with PlotGeneGO
}
\description{
Compute function of gene using GO over hitting probablities, i.e, which genes are regulated
by this gene?
}
