% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{ListOfNandoNetwork-class}
\alias{ListOfNandoNetwork-class}
\alias{ListOfNandoNetwork}
\title{An S4 class to represent a list of networks}
\description{
An S4 class to represent a list of networks
}
\section{Slots}{

\describe{
\item{\code{nets}}{All networks}
}}

